requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
chromadb==0.3.23
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==0.8.11
mammoth==1.6.0
aiofiles==23.1.0
